extract device_data.rs and robot_data.rs
