+++ src/model/library_downloader.rs
@@ -1,11 +1,12 @@
 // Copyright (c) 2021 Marco Boneberger
 // Licensed under the EUPL-1.2-or-later
 use crate::exception::FrankaException::ModelException;
-use crate::network::{Network, RobotData};
+use crate::network::Network;
 use crate::robot::service_types::{
     LoadModelLibraryArchitecture, LoadModelLibraryRequest, LoadModelLibraryStatus,
     LoadModelLibrarySystem,
 };
+use crate::robot_data::RobotData;
 use crate::FrankaResult;
 use std::fmt;
 use std::fmt::Display;
